Handle file renames in a Pull Request
Problem: When files were renamed as part of the series of commits on a pull request, the checker would attempt to look at a file at a given point that no longer existed and would fail, then the exception would not be logged to the Pull Request.

Solution: Keep a list of the renamed files, and remove them from the built list of items to check before validating encoding. Also, make sure that if we do encounter an exception, that the message is logged to the pull request as a new comment thread.